docs(common): Document internal form component for the login page
Refs: #272
